Fix support for Ctrl-C on the macosx backend.
Support is largely copy-pasted from, and tests are shared with, the qt
implementation (qt_compat._maybe_allow_interrupt), the main difference
being that what we need from QSocketNotifier, as well as the equivalent
for QApplication.quit(), are reimplemented in ObjC.

qt_compat._maybe_allow_interrupt is also slightly cleaned up by moving
out the "do-nothing" case (`old_sigint_handler in (None, SIG_IGN, SIG_DFL)`)
and dedenting the rest, instead of keeping track of whether signals were
actually manipulated via a `skip` variable.

Factoring out the common parts of _maybe_allow_interrupt is left as a
follow-up.

(Test e.g. with
`MPLBACKEND=macosx python -c "from pylab import *; plot(); show()"`
followed by Ctrl-C.)
